+++ aider/__init__.py
@@ -1,6 +1,6 @@
 from packaging import version
 
-__version__ = "0.77.1.dev"
+__version__ = "0.78.1.dev"
 safe_version = __version__
 
 try: